# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "miIFTS API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/registro": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "description": "Registra un nuevo usuario.",
        "operationId": "register_auth_registro_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "description": "Inicia sesión y devuelve un token JWT.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "description": "Devuelve los datos del usuario autenticado (protegido).",
        "operationId": "get_me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Update Me",
        "description": "Actualiza el nombre y/o apellido del usuario autenticado.",
        "operationId": "update_me_auth_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/auth/verify": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Token",
        "operationId": "verify_token_auth_verify_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Forgot Password",
        "description": "Dispara el envío del email de reset. Responde siempre lo mismo,\nexista o no el email, para no revelar qué direcciones están registradas.",
        "operationId": "forgot_password_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset Password",
        "description": "Consume el token de un solo uso y actualiza la contraseña.",
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/password": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Cambiar Password",
        "description": "Cambia la contraseña del usuario autenticado. Requiere la actual.",
        "operationId": "cambiar_password_auth_password_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CambiarPasswordRequest": {
        "properties": {
          "password_actual": {
            "type": "string",
            "title": "Password Actual",
            "examples": [
              "secreta123"
            ]
          },
          "password_nueva": {
            "type": "string",
            "minLength": 8,
            "title": "Password Nueva",
            "description": "Contraseña nueva (mínimo 8 caracteres, al menos una letra y un número)",
            "examples": [
              "nuevaSecreta123"
            ]
          }
        },
        "type": "object",
        "required": [
          "password_actual",
          "password_nueva"
        ],
        "title": "CambiarPasswordRequest",
        "description": "`PATCH /auth/password` — el usuario ya está logueado, pero igual tiene\nque reprobar identidad con la contraseña actual (si alguien le roba la\nsesión, no puede usarla para expulsarlo cambiándole la contraseña)."
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "ada@ifts.edu.ar"
            ]
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "description": "`POST /auth/forgot-password` — dispara el envío del email de reset."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MensajeResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "MensajeResponse",
        "description": "Respuesta genérica de un solo mensaje (forgot/reset password)."
      },
...
      "PerfilUpdate": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre",
            "examples": [
              "Ada"
            ]
          },
          "apellido": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Apellido",
            "examples": [
              "Lovelace"
            ]
          }
        },
        "type": "object",
        "title": "PerfilUpdate",
        "description": "`PATCH /auth/me` — el alumno sólo edita su propio nombre y apellido.\n\nLa carrera se muestra en el perfil pero no se cambia desde acá. El email\nidentifica la cuenta y el rol nunca se acepta desde el request. Cualquier\notro campo del body se ignora."
      },
...
      "ResetPasswordRequest": {
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "title": "Token",
            "examples": [
              "b8f2b3b7c1a94e6c9d0e..."
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Contraseña nueva (mínimo 8 caracteres, al menos una letra y un número)",
            "examples": [
              "nuevaSecreta123"
            ]
          }
        },
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "title": "ResetPasswordRequest",
        "description": "`POST /auth/reset-password` — token recibido por email + contraseña nueva.\n\nMisma validación de contraseña que `UsuarioCreate`."
      },
      "RolUsuario": {
        "type": "string",
        "enum": [
          "estudiante",
          "admin"
        ],
        "title": "RolUsuario"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "usuario": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UsuarioResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "description": "Respuesta del endpoint de login."
      },
      "UsuarioCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Nombre",
            "description": "Nombre",
            "examples": [
              "Ada"
            ]
          },
          "apellido": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Apellido",
            "description": "Apellido",
            "examples": [
              "Lovelace"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Correo electrónico único",
            "examples": [
              "ada@ifts.edu.ar"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Contraseña en texto plano (mínimo 8 caracteres, al menos una letra y un número)",
            "examples": [
              "secreta123"
            ]
          },
          "carrera_id": {
            "type": "integer",
            "title": "Carrera Id",
            "description": "ID de la carrera a la que pertenece",
            "examples": [
              1
            ]
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "apellido",
          "email",
          "password",
          "carrera_id"
        ],
        "title": "UsuarioCreate",
        "description": "Datos necesarios para registrar un nuevo usuario.\n\nNota de seguridad: el rol NO se acepta desde el request. El registro\npúblico siempre crea usuarios con rol `estudiante`; los administradores\nse crean por seed o promoción manual en la base de datos."
      },
      "UsuarioLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "ada@ifts.edu.ar"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "examples": [
              "secreta123"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UsuarioLogin",
        "description": "Datos necesarios para iniciar sesión."
      },
      "UsuarioResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "carrera_id": {
            "type": "integer",
            "title": "Carrera Id"
          },
          "fecha_registro": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Registro"
          },
          "rol": {
            "$ref": "#/components/schemas/RolUsuario"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "apellido",
          "email",
          "carrera_id",
          "fecha_registro",
          "rol"
        ],
        "title": "UsuarioResponse",
        "description": "Datos del usuario que se devuelven al frontend (sin password_hash)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "user_id"
        ],
        "title": "VerifyResponse",
        "description": "Respuesta de `GET /auth/verify`."
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}